changed error message and separated components
